storage_service/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confluent-kafka==2.4.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
distlib==0.3.7
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.12.2
google==3.0.0
greenlet==3.0.3
grpcio==1.58.0
grpcio-tools==1.58.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
mysqlclient==2.2.4
pid==3.0.4
platformdirs==3.10.0
prompt_toolkit==3.0.45
protobuf==4.24.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
redis==5.0.4
requests==2.32.3
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.24.6
wcwidth==0.2.13
websockets==11.0.3
zope.interface==6.4.post2
